Move change and update listeners for special fields
