+++ src/main/java/net/sf/jabref/gui/specialfields/SpecialFieldUpdateListener.java
@@ -1,4 +1,4 @@
-package net.sf.jabref.specialfields;
+package net.sf.jabref.gui.specialfields;
 
 import javax.swing.SwingUtilities;
 
@@ -6,6 +6,7 @@ import net.sf.jabref.JabRefGUI;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.model.entry.FieldName;
 import net.sf.jabref.model.entry.event.FieldChangedEvent;
+import net.sf.jabref.specialfields.SpecialFieldsUtils;
 
 import com.google.common.eventbus.Subscribe;
